requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
